+++ frontend/src/components/Explore/Explore.module.css
@@ -1,14 +1,3 @@
-.buttons button{
-    color: black;
-    margin-right: .5vw;
-    cursor: pointer;
-    transition: background-color 0.1s;
-}
-
-.buttons button:hover{
-    background-color: #f76359;
-}
-
 .postsContainer {
     display: grid;
     grid-row-start: auto;
@@ -39,6 +28,3 @@
     background-color: #0a0a0a;
 }
 
-.current {
-    background-color: #f76359;
-}
