+++ frontend/src/utils/FilterButtons/FilterButtons.js
@@ -1,10 +1,9 @@
-import styles from "./Explore.module.css";
+import styles from "./FilterButtons.module.css";
 import { useState } from "react";
-import api from "../../utils/api";
-import { getExploreListUrl } from "@/utils/url-segments";
+import api from "../api";
 import set_hot_score from "./set-hot-score.js";
 
-const FilterButtons = ({posts, setPosts}) => {
+const FilterButtons = ({posts, setPosts, postsUrl}) => {
     const [time, setTime] = useState('week');
     const [sort, setSort] = useState('top'); 
     const [currentTimeButton, setCurrentTimeButton] = useState('week');
@@ -14,7 +13,7 @@ const FilterButtons = ({posts, setPosts}) => {
         if(time === newTime) return;
         console.log("Fetching explore page posts...");
         try {
-            const response = await api.get(getExploreListUrl(newTime, sort));
+            const response = await api.get(postsUrl(newTime, sort));
             setTime(newTime);
             setPosts(response.data);
             console.log(response.data);
